chore(client): rename `Invite.svelte` => `AddInvite.svelte`
